qtopia-opensource-src-4.2.2

+++ scripts/greenphone-make-flash.sh
@@ -151,9 +151,9 @@ install_extras()
     sudo chown -R root.root $MNTPNT/usr/share/zoneinfo
 }
 
-partition_rootfs()
+partition_tffsa()
 {
-    echo "Partitioning rootfs"
+    echo "Partitioning disk1 image"
 
     ROOTFS_PARTITION=`mktemp rootfs_part.XXXXXX`
     QTOPIA_PARTITION=fimage/qtopia.cramfs
@@ -175,11 +175,11 @@ partition_rootfs()
         exit 1
     fi
 
-    dd if=/dev/zero of=$ROOTFS_FILENAME bs=1 count=0 seek=$(($ROOTFS_DEVICE_SIZE * 512)) 2>/dev/null
+    dd if=/dev/zero of=$DISK1_FILENAME bs=1 count=0 seek=$(($ROOTFS_DEVICE_SIZE * 512)) 2>/dev/null
 
     for i in 0 1 2 3 4 5 6 7; do
         LOOP_DEVICE=/dev/loop$i
-        sudo /sbin/losetup $LOOP_DEVICE $ROOTFS_FILENAME 2>/dev/null && break
+        sudo /sbin/losetup $LOOP_DEVICE $DISK1_FILENAME 2>/dev/null && break
     done
 
     # create partitions
@@ -192,8 +192,8 @@ EOF
 
     sudo /sbin/losetup -d $LOOP_DEVICE
 
-    dd if=$ROOTFS_PARTITION of=$ROOTFS_FILENAME bs=512 seek=$FIRST_SECTOR conv=notrunc 2>/dev/null
-    dd if=$QTOPIA_PARTITION of=$ROOTFS_FILENAME bs=512 seek=$(($FIRST_SECTOR + $ROOTFS_PARTITION_SIZE)) conv=notrunc 2>/dev/null
+    dd if=$ROOTFS_PARTITION of=$DISK1_FILENAME bs=512 seek=$FIRST_SECTOR conv=notrunc 2>/dev/null
+    dd if=$QTOPIA_PARTITION of=$DISK1_FILENAME bs=512 seek=$(($FIRST_SECTOR + $ROOTFS_PARTITION_SIZE)) conv=notrunc 2>/dev/null
 
     rm -f $ROOTFS_PARTITION
 }
@@ -201,7 +201,13 @@ EOF
 make_flash_image()
 {
     echo "Creating USB Flash image"
+
+    local OUTPUT_IMAGE="$1"
+    shift
+
     rm -f flash.conf
+
+    IMAGE_FILES=""
 
     if [ ! -z $KERNEL_IMAGE ] && [ ! -z $KERNEL_FILENAME ] && [ -f $KERNEL_FILENAME ]; then
         MD5SUM=`md5sum $KERNEL_FILENAME | awk '{print $1}'`
@@ -209,14 +215,39 @@ make_flash_image()
         IMAGE_FILES="$IMAGE_FILES $KERNEL_FILENAME"
     fi
 
-    if [ ! -z $ROOTFS_IMAGE ] && [ ! -z $ROOTFS_FILENAME ]; then
-        MD5SUM=`md5sum $ROOTFS_FILENAME | awk '{print $1}'`
-        echo "$ROOTFS_NAME ($ROOTFS_VERSION):$ROOTFS_FILENAME:3:$MD5SUM" >> flash.conf
-        IMAGE_FILES="$IMAGE_FILES $ROOTFS_FILENAME"
+    if [ ! -z $ROOTFS_IMAGE ] && [ ! -z $DISK1_FILENAME ]; then
+        MD5SUM=`md5sum $DISK1_FILENAME | awk '{print $1}'`
+        echo "$DISK1_NAME ($ROOTFS_VERSION):$DISK1_FILENAME:3:$MD5SUM" >> flash.conf
+        IMAGE_FILES="$IMAGE_FILES $DISK1_FILENAME"
     fi
 
-    tar --remove-files -czhf qtopia-greenphone-flash flash.conf $IMAGE_FILES
-    echo "Finished creating USB Flash image : $PWD/qtopia-greenphone-flash"
+    while [ $# -gt 0 ]; do
+        case $1 in
+            --blank-disk2)
+                dd if=/dev/zero of=$DISK2_FILENAME bs=1048576 count=32 2>/dev/null
+                MD5SUM=`md5sum $DISK2_FILENAME | awk '{print $1}'`
+                echo "$DISK2_NAME:$DISK2_FILENAME:4:$MD5SUM" >> flash.conf
+                IMAGE_FILES="$IMAGE_FILES $DISK2_FILENAME"
+                ;;
+            --blank-user)
+                dd if=/dev/zero of=$USER_FILENAME bs=1048576 count=5 2>/dev/null
+                MD5SUM=`md5sum $USER_FILENAME | awk '{print $1}'`
+                echo "$USER_NAME:$USER_FILENAME:5:$MD5SUM" >> flash.conf
+                IMAGE_FILES="$IMAGE_FILES $USER_FILENAME"
+                ;;
+            --blank-doc)
+                dd if=/dev/zero of=$DOC_FILENAME bs=1024 count=36480 2>/dev/null
+                MD5SUM=`md5sum $DOC_FILENAME | awk '{print $1}'`
+                echo "$DOC_NAME:$DOC_FILENAME:6:$MD5SUM" >> flash.conf
+                IMAGE_FILES="$IMAGE_FILES $DOC_FILENAME"
+                ;;
+        esac
+
+        shift
+    done
+
+    tar -czhf "$OUTPUT_IMAGE" flash.conf $IMAGE_FILES
+    echo "Finished creating USB Flash image : $PWD/$OUTPUT_IMAGE"
 }
 
 make_updateqtopia_image()
@@ -227,16 +258,19 @@ make_updateqtopia_image()
     mkdir fimage
     $QTOPIA_SOURCE_PATH/scripts/greenphone-make-cramfs.sh $QTOPIA_IMAGE_PATH fimage
 
-    cp $QTOPIA_SOURCE_PATH/devices/greenphone/src/devtools/flash-files/*.gif fimage
+    cp $QTOPIA_SOURCE_PATH/devices/greenphone/src/devtools/flash-files/flash-status-*.gif fimage
     chmod 660 fimage/*.gif
     cp $QTOPIA_SOURCE_PATH/devices/greenphone/src/devtools/flash-files/trolltech_flash.sh fimage
     chmod 775 fimage/trolltech_flash.sh
 
     if [ -s fimage/trolltech_flash.sh -a -s fimage/qtopia.cramfs -a -s fimage/flash-status-flashing.gif ]; then
-        tar -czhf qtopia-greenphone-update.tar.gz -C fimage .
-        echo "Finished creating updateqtopia image : $PWD/qtopia-greenphone-update.tar.gz"
+        tar -czhf $UPDATE_IMAGE.tar.gz -C fimage .
+        echo "Finished creating updateqtopia image : $PWD/$UPDATE_IMAGE.tar.gz"
+        zip -jq $UPDATE_IMAGE.zip fimage/*
+        echo "Finished creating updateqtopia image : $PWD/$UPDATE_IMAGE.zip"
     else
-        rm -f qtopia-greenphone-update.tar.gz # clean up after ourselves
+        rm -f $UPDATE_IMAGE.tar.gz # clean up after ourselves
+        rm -f $UPDATE_IMAGE.zip
         echo "Error: Failed to create updateqtopia image"
         exit 1
     fi
@@ -245,7 +279,7 @@ make_updateqtopia_image()
 usage()
 {
     echo -e "Usage: `basename $0` [--clean] [--qtopia] [--flash] [--optimize] [--qtopia-build <path>] [--qtopia-source <path>] [--rootfs <file>]"
-    echo -e "If no options are specified defaults to --qtopia --qtopia-build \$PWD/image/opt/Qtopia --qtopia-source \$PWD --rootfs \$PWD/rootfs/greenphone_rootfs.ext2 --kernel \$PWD/rootfs/greenphone_kernel\n" \
+    echo -e "If no options are specified defaults to --qtopia --qtopia-build \$PWD/image --qtopia-source \$PWD --rootfs \$PWD/rootfs/greenphone_rootfs.ext2 --kernel \$PWD/rootfs/greenphone_kernel\n" \
             "   --clean          Clean image files.\n" \
             "   --qtopia         Make Qtopia image.\n" \
             "   --flash          Make flash image.\n" \
@@ -253,6 +287,7 @@ usage()
             "   --qtopia-source  Location of Qtopia source tree.\n" \
             "   --rootfs         Location of rootfs image.\n" \
             "   --kernel         Location of Linux kernel image.\n" \
+            "   --build-path     Location of rootfs build directory.\n" \
             "   --optimize       Optimize dynamic shared objects. Requires network access to a Greenphone.\n" \
             "                    Greenphone IP address can be specified with the GREENPHONE_IP environment variable.\n"
 }
@@ -264,18 +299,35 @@ OPTION_OPTIMIZE=0
 OPTION_QTOPIA_IMAGE=0
 OPTION_FLASH_IMAGE=0
 
+BUILD_PATH=$PWD/rootfs
+ 
 DEFAULT_IMAGES=1
 
-QTOPIA_SOURCE_PATH=$PWD
-QTOPIA_BUILD_PATH=$PWD/image/opt/Qtopia
+if [ -z $QTOPIA_DEPOT_PATH ]; then
+		QTOPIA_SOURCE_PATH=$PWD
+else
+		QTOPIA_SOURCE_PATH=$QTOPIA_DEPOT_PATH
+fi
+
+QTOPIA_BUILD_PATH=$PWD/image
 
 # Default Flash image data
-KERNEL_NAME="Trolltech Greenphone Kernel (2.4.19)"
+KERNEL_NAME="Trolltech Greenphone Kernel (2.4.19 w/ LIDS)"
 KERNEL_FILENAME=zImage_Truly_Trolltech
 KERNEL_SIZELIMIT=1048576
-ROOTFS_NAME="Trolltech Greenphone Root Filesystem"
+DISK1_NAME="Trolltech Greenphone Root and Qtopia Filesystems"
 ROOTFS_VERSION="unknown"
-ROOTFS_FILENAME=greenphone_root_prg_FS.ext2
+DISK1_FILENAME=greenphone_root_qtopia_FS.bin
+DISK2_NAME="Blank Disk2"
+DISK2_FILENAME=greenphone_disk2_FS.ext2
+USER_NAME="Blank User"
+USER_FILENAME=greenphone_user_FS.ext2
+DOC_NAME="Blank Documents"
+DOC_FILENAME=greenphone_doc_FS.vfat
+
+# Output filenames
+FLASH_IMAGE="qtopia-greenphone-flash"
+UPDATE_IMAGE="qtopia-greenphone-update"
 
 while [ $# -ne 0 ]; do
     case $1 in
@@ -337,6 +389,10 @@ while [ $# -ne 0 ]; do
             OPTION_OPTIMIZE=1
             DEFAULT_OPTIONS=0
             ;;
+        --build-path)
+            BUILD_PATH="$2"
+            shift 1
+            ;;
         --help)
             usage
             exit 0
@@ -355,19 +411,46 @@ if [ $DEFAULT_OPTIONS -eq 1 ]; then
 fi
 
 if [ $DEFAULT_IMAGES -eq 1 ]; then
-    ROOTFS_IMAGE=$PWD/rootfs/greenphone_rootfs.ext2
-    KERNEL_IMAGE=$PWD/rootfs/greenphone_kernel
+    ROOTFS_IMAGE=$BUILD_PATH/greenphone_rootfs.ext2
+    KERNEL_IMAGE=$BUILD_PATH/greenphone_kernel
 fi
+# Make sure all input files and directories exist
+[ -d "$QTOPIA_BUILD_PATH" ] || die "$QTOPIA_BUILD_PATH does not exist or is not a directory"
+if [ ! -f "$QTOPIA_BUILD_PATH/qpe.sh" ]; then
+ cp -a $QTOPIA_DEPOT_PATH/devices/greenphone/src/devtools/startup/qpe.sh $QTOPIA_BUILD_PATH
+ cp -a $QTOPIA_DEPOT_PATH/devices/greenphone/src/devtools/startup/qpe.env $QTOPIA_BUILD_PATH
+fi
+if [ ! -f "$QTOPIA_BUILD_PATH/qpe.sh" ]; then
+    echo "$QTOPIA_BUILD_PATH does not contain an image - check \"--qtopia-build\" argument"
+    if [ -f "$QTOPIA_BUILD_PATH/config.status" ]; then
+        echo "    $QTOPIA_BUILD_PATH looks like the build tree"
+        QTOPIA_BUILD_PATH="$QTOPIA_BUILD_PATH/image"
+        if [ -f "$QTOPIA_BUILD_PATH/qpe.sh" ]; then
+            echo "    using $QTOPIA_BUILD_PATH as image"
+        else
+            exit 1;
+        fi
+    else
+        exit 1;
+    fi
+fi
+
+[ -d "$QTOPIA_SOURCE_PATH" ] || die "Qtopia source path does not exist or is not a directory"
 
 if [ $OPTION_CLEAN -eq 1 ]; then
     rm -rf fimage
-    rm -rf $KERNEL_FILENAME
-    rm -rf $ROOTFS_FILENAME
-    rm -rf qtopia_greenphone
-    rm -rf qtopia_greenphone_flash
+    rm -f $KERNEL_FILENAME
+    rm -f $DISK1_FILENAME
+    rm -f $UPDATE_IMAGE.tar.gz
+    rm -f $UPDATE_IMAGE.zip
+    rm -f $FLASH_IMAGE
+    rm -f $FLASH_IMAGE-full
 fi
 
 if [ $OPTION_FLASH_IMAGE -eq 1 ]; then
+    if [ ! -z "$KERNEL_IMAGE" ]; then
+        [ -f "$KERNEL_IMAGE" ] || die "Kernel image does not exist"
+    fi
     if [ ! -z $KERNEL_IMAGE ]; then
         KERNEL_SIZE=`stat -c '%s' $KERNEL_IMAGE`
 
@@ -377,14 +460,17 @@ if [ $OPTION_FLASH_IMAGE -eq 1 ]; then
         cp $KERNEL_IMAGE $KERNEL_FILENAME
     fi
 
+    if [ ! -z "$ROOTFS_IMAGE" ]; then
+        [ -f "$ROOTFS_IMAGE" ] || die "Rootfs image $ROOTFS_IMAGE does not exist"
+    fi
     if [ ! -z $ROOTFS_IMAGE ]; then
         OPTION_QTOPIA_IMAGE=1
 
-        cp $ROOTFS_IMAGE $ROOTFS_FILENAME
+        cp $ROOTFS_IMAGE $DISK1_FILENAME
 
         MNTPNT=`mktemp -d flash.XXXXXX`
         [ -d $MNTPNT ] || die "Could not create temporary rootfs directory"
-        sudo mount -o loop,rw $ROOTFS_FILENAME $MNTPNT
+        sudo mount -o loop,rw $DISK1_FILENAME $MNTPNT
 
         if [ -r $MNTPNT/fs.ver ]; then
             ROOTFS_VERSION=`sed -ne '/^troll/ s/^troll v//p' < $MNTPNT/fs.ver`
@@ -417,13 +503,16 @@ if [ $OPTION_QTOPIA_IMAGE -eq 1 ]; then
 fi
 
 if [ $OPTION_FLASH_IMAGE -eq 1 ] && [ ! -z $ROOTFS_IMAGE ]; then
-    partition_rootfs
+    partition_tffsa
 
     rmdir $MNTPNT
 fi
 
 if [ $OPTION_FLASH_IMAGE -eq 1 ]; then
-    make_flash_image
+    make_flash_image $FLASH_IMAGE
+    make_flash_image $FLASH_IMAGE-full --blank-disk2 --blank-user --blank-doc
+
+    rm -f $KERNEL_FILENAME $DISK1_FILENAME $DISK2_FILENAME $USER_FILENAME $DOC_FILENAME
 fi
